BackEnd+PaiementStripe

+++ src/main/java/tn/esprit/ecocycletech/Service/AppareilsManagement/AppareilServiceImpl.java
@@ -24,9 +24,13 @@ public class AppareilServiceImpl implements IAppareilService {
 
     @Override
     public boolean deleteAppareil(int id) {
-        appareilRepository.deleteById(id);
+        if (appareilRepository.existsById(id)) {
+            appareilRepository.deleteById(id);
+            return true;
+        }
         return false;
     }
+
 
     // Récupérer tous les appareils
     @Override
@@ -34,7 +38,6 @@ public class AppareilServiceImpl implements IAppareilService {
         return appareilRepository.findAll();
     }
 
-    // Mettre à jour un appareil existant
     @Override
     public Appareil updateAppareil(int id, Appareil appareil) {
         Optional<Appareil> existingAppareilOpt = appareilRepository.findById(id);
@@ -44,7 +47,6 @@ public class AppareilServiceImpl implements IAppareilService {
             existingAppareil.setCategorie(appareil.getCategorie());
             existingAppareil.setEtatAppareil(appareil.getEtatAppareil());
             existingAppareil.setMarque(appareil.getMarque());
-            existingAppareil.setQuantite(appareil.getQuantite());
             existingAppareil.setPrix(appareil.getPrix());
             existingAppareil.setDescription(appareil.getDescription());
             existingAppareil.setImageurl(appareil.getImageurl());
